+++ packages/flzl2008/index.html
@@ -11,9 +11,7 @@
 </head>
 
 <body>
-  <div id="app" class="px-4">
-    
-  </div>
+  <div id="app" class="app px-4"></div>
   <script type="module" src="./dist/index.js"></script>
 </body>
