+++ packages/flzl2008/index.ts
@@ -1,4 +1,4 @@
 import CafeMenuApp from './src/js/components/cafeMenuApp.js';
 import $ from './src/js/utils/commons.js';
 
-new CafeMenuApp($('#app'));
+new CafeMenuApp($('.app'));
